requirements.txt
gtts>=2.5.0
faster-whisper>=0.10.0
pydub>=0.25.0 
librosa>=0.10.0
fastapi
uvicorn[standard]
pydantic
pydub
google-generativeai
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
google-genai
langchain-google-genai
langchain
langchain-community
langchain-qdrant
qdrant-client >=1.16.2
fastembed
python-multipart
aiofiles
librosa
soundfile


streamlit
supabase
google-generativeai
pandas